Add ability to store user active room, with skeleton accounts.  - Adds a user_state table, currently only with active_room.  - A user must have an account to take advantage of state.  - Now, all users will get an 'account' even if they don't explicitly register.  - Bonus: converts user queries to compile-time checked macros.
To support these automatically created "accounts," the accounts table
now also has an account_status column, indicating if the user is
registered or not (or pending activation--future use).

The User model has been updated with extra properties from the state,
and the user is now carrried in the Context during command execution.
A user is ensured to be created before executing the command.